+++ git-config/tests/integration_tests/value_integeration_tests.rs
@@ -1,12 +1,6 @@
-use std::{borrow::Cow, convert::TryFrom, path::Path};
+use std::{borrow::Cow, convert::TryFrom};
 
 use git_config::{file::GitConfig, values::*};
-
-#[test]
-fn parse_config_with_windows_line_endings_successfully() -> crate::Result {
-    GitConfig::open(Path::new("tests").join("fixtures").join("repo-config.crlf"))?;
-    Ok(())
-}
 
 /// Asserts we can cast into all variants of our type
 #[test]
